% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_trees_dist_forcebin_b.r
\name{genome.trees.dist.forcebin.b}
\alias{genome.trees.dist.forcebin.b}
\title{Use binary tree metrics to compare two inferred ancestries (ARGs) over a genomic region}
\usage{
genome.trees.dist.forcebin.b(treeseq.a, treeseq.b, replicates = 1,
  seed = NA, acceptable.length.diff.pct = 0.1, variant.positions = NULL,
  threads = 1)
}
\arguments{
\item{treeseq.a}{The base multiPhylo object, or a path to a .nex file}

\item{treeseq.b}{The multiPhylo object (or path to .nex files) containing a nonbinary tree 
to compare to treeseq.a.}

\item{replicates}{The number of times to run genome.trees.dist()}

\item{seed}{The random seed to give to set.seed before starting random polytomy breaking}

\item{acceptable.length.diff.pct}{How much difference in sequence length is allowed between the 2 trees? (Default: 0.1 percent)}

\item{variant.positions}{A list of positions of each variant (not implemented)}
}
\value{
an average of the metrics over each of the replicates
}
\description{
Runs genome.trees.dist(..., randomly.resolve.polytomies=TRUE) multiple times on the same
two tree sequences, breaking polytomies in the second tree sequence differentially (at random)
each time.
See ?genome.trees.dist for more information.
}
\examples{
genome.trees.dist.forcebin()
}
