% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_trees_dist_multi.r
\name{genome.trees.dist.multi}
\alias{genome.trees.dist.multi}
\title{Use tree metrics to compare multiple inferred ancestries (ARGs) over a genomic region}
\usage{
genome.trees.dist.multi(treeseq.base, treeseq.multi,
  acceptable.length.diff.pct = 0.1, variant.positions = NULL,
  randomly.resolve.polytomies = FALSE)
}
\arguments{
\item{treeseq.base}{The base multiPhylo object, or a path to a .nex file}

\item{treeseq.multi}{A list of multiPhylo objects (or list of paths to .nex files).
If the list has names which can be all be converted to numbers, then these are treated as 
weights, and the returned value is a single average set of distance measures over all
the different tree sequences.}

\item{acceptable.length.diff.pct}{How much difference in sequence length is allows between the 2 trees? (Default: 0.1 percent)}

\item{variant.positions}{A list of positions of each variant (not implemented)}

\item{randomly.resolve.polytomies}{Some distance metrics only operate on binary trees. Set this to TRUE to force trees to be binary
by randomly resolving polytomies where necessary. If a number, it is passed to set.seed as a RNG seed.}
}
\description{
Runs genome.trees.dist to compare multiple ancestry estimates against a known original
See ?genome.trees.dist for more information.
}
\examples{
genome.trees.dist.multi()
}
