% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARGmetrics.r
\name{genome.trees.dist}
\alias{genome.trees.dist}
\title{Compare multiple trees across a genomic region}
\usage{
genome.trees.dist(a, b, output.full.table = FALSE,
  acceptable.length.diff.pct = 0.1, variant.positions = NULL)
}
\arguments{
\item{a}{The first multiPhylo object}

\item{b}{The second multiPhylo object}

\item{output.full.table}{Output tree metrics for each overlapping region, rather than simply a weighted summary.}

\item{acceptable.length.diff.pct}{How much difference in sequence length is allows between the 2 trees? (Default: 0.1%)}

\item{variant.positions}{A list of positions of each variant (not implemented)}
}
\description{
Requires multiPhylo objects representing sequential trees along a genomic region.
The tree names should be numbers, which represent the rightmost position along a
genome (non-inclusive) covered by each tree.
}
\examples{
genome.trees.dist()
}
