% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_read_nexus.r
\name{new.read.nexus}
\alias{new.read.nexus}
\title{Replacement for read.nexus in the ape library}
\usage{
new.read.nexus(file, tree.names = NULL, force.multi = FALSE)
}
\arguments{
\item{file}{A path to the nexus file.}

\item{tree.names}{Override or provide names for the trees in the file.}

\item{force.multi}{Return a multiPhylo object, even if there is only one tree in the file.}
}
\description{
This should be identical to read.nexus, but with a parameter force.multi
to force the return of a multiPhylo object, even if there is only one
tree in the nexus file
}
\examples{
See ?read.nexus
}
